+++ index.ts
@@ -1,1 +1,5 @@
 export * from './src';
+
+import * as iap from './src/iap';
+
+export default iap;
